Extract setting game dir from NavigationMenu.prepareLaunch
The main purpose is to make the functionality for ensuring a game dir
is set reusable in other places.

While refactoring, the implementation was changed so that the game dir
is no longer set if it was set already. I'm not 100% certain I read the
code correctly, but it seems to me the old implementation would only set
the current value again, which is a no-op.

Refs TS-530