Extract launch functionality from NavigationMenu
The main purpose is to make the functionality reusable in other places,
but this also DRYs up the NavigationMenu quite nicely.

resolveGameDir checks two separate directories, which seems a bit odd
but I'm so unfamiliar with the project that there's probably a good
reason for it and I've left that as it was.

Refs TS-530