Add support for syncing mods via custom scheme link
This is a MVP implementation that leaves much to hope for.

Currently actions triggered by the links are defined on component-basis.
This means use needs to have a correct page already open in the mod
manager for the link to actually do what it's supposed to. Ideally link
actions would be defined globally and would support moving to the
correct page if necessary.

Both current implementations (installing a mod and syncing mods from a
server listing) also require the mod manager to have the correct game
and suitable mod profile selected for the actions to work. User isn't
even notified about the issue: the links simply do nothing.

Current actions are also tightly coupled with the component they're
defined in. This could perhaps be solved with a provider-based approach,
which would also allow defining the actions centrally in one place,
reducing duplication and avoiding conflicts.

Error handling is currently quite non-existent.

Both links are currently initialized with old hookModInstallProtocol,
which is now a misnomer. Renaming it would also require changes to
non-Thunderstore Mod Manager parts, and was left out the scope of this
commit.

Refs TS-682