Refactor FolderMigration
Attempt to improve code quality, although most of the changes are
subjective.

Move the migration check into Vuex store so it doesn't get called
multiple times now that the game selection screen can be accessed
without restarting the app.

The cache cleaning was removed, since it failed if the old directory
structure was still used: it tried to read the profiles from a non-
existing dir.

Refs TS-1138