Add standalone GameRunningModal
The modal was originally built in the navigation menu since that's
where the games are launched from, but decoupling these two will
hopefully make it easier to customize the NavigationMenu in different
views and mod manager variants.

Since the modal and the NavigationMenu are no longer a direct parent-
child relation, the status of modal visibity is stored in Vuex store.
A separate store module was created for this purpose - the idea is that
in the future it could be used to manage more modals than just the one
it currently does.