Move mod filters to Vuex store
Some of the filters in manager page are managed via a modal. Having the
modal inside the Manager page makes it difficult to move towards route
based navigation. But before the modal can be moved out of the manager
page, we need to move the management of filter state to make them
universally accessible.

Refs TS-1309