Add standalone CategoryFilterModal
Continue making Manager page less of a god-class component.

This also partially fixes the problem where "allow NSFW" checkbox
didn't react correctly to dark/light theme. The value is now read from
correct place, so the checkbox is themed based on the theme that was
active when user entered Manager page. However, if user changes the
theme while on manager page, the checkbox theme is not updated yet. I
wasn't sure how this should be solved, and it falls outside the scope
of the ticket, so I just marked the issue with TODO comment.

Refs TS-1309