Change online mod list filtering
Instead of passing the filtering function as a prop to the modal that
handles updating filter values, watch changes to those values instead.
This makes it easier to move the filtering function to another module
in the future.

As a side effect of this change, the mod list is now updated
immediately when a filter is changed. Keen eyes can detect the mod list
updating under the modal. I didn't detect any performance issues with
this approach, but if problems arises with this approach, we can change
the modal to store the changes to filters locally and commit them to
Vuex store only when the modal is closed.

Instead of calling the method responsible for sorting the mod list
(which calls the filtering method behind the scenes), call the
filtering method directly. Changes done in the filter modal don't
affect the sorting, so sorting the list seems unnecessary.

Refs TS-1311