Create a standalone component from Manager page's "online" subview
The idea is that the selected subview will later be selected
automatically by the routing component, but for this to work the view
pretty much needs to be a single component.

Since the component will be defined in the routing config in the future
it made sense to me that the component is imported directly rather than
via a provider. To my understanding the point of the providers is to
allow swapping components for customization, and in the future that
will be handled by the router config.

Refs TS-1311