Simplify theme change detection
Earlier the subviews of the Manager page were all mounted at the same
time, and the other two needed the poll the settings object to detect
changes done in the settings subview. Now only one of the views is
mounted at the time, so it's enough to check the settings when a
subview is mounted.