Ensure settings can be loaded before mounting the app
For example a corrupted IndexedDB storage may cause loading the
settings to fail, which would result in the users seeing an emptyish
screen. To prevent that, try to load the settings as the first thing
when the app is launched. If that fails, log the error for debug
purposes and display an error modal to the user.

Refs TS-1357