Add support for forcing proton treatment for games
Make it possible to force the manager to treat the game as a proton game
by including a file with a specific name in the game directory.

This is not intended as a full solution, but rather a workaround which
doesn't require editing of steam's config files to align with what the
manager expects them to be for proton games.

Resolves #950