Add support for several new games
Add support for the following games:
- Brotato
- Ancient Dungeon VR
- Atrio: The Dark Wild
- Ultimate Chicken Horse

Additionally, add support for a new install rule type which simply moves
the package zip to the target directory, as that's what the godot mod
loader expects.

This commit does not yet include installation logic for the godot mod
loader, as that doesn't have a locked in spec yet.