Sort the profile list
Ensure the profile list is alphabetically sorted. This has been
happening coincidentally in scenarios where the I/O promises resolve in
order, but this might not always be guaranteed.

This isn't the most performant way to handle it (as the list is sorted
after every new entry is added to it), but at the scales we're operating
with, it shouldn't matter.