Make modloader mappings global
Edit the hardcoded modloader mappings so that any modloader package can
be used with any game.

Rationale:

There has been a few instances in the past where a community has swapped
the package they use for mod loaders, and until now it has always
required an update to the hardcoded definitions. While this change does
not yet remove the hardcoding entirely, it brings the system more in
line with the future goals of having the packages declare how they
should be installed themselves. In other words, whether or not a package
is handled as a mod loader should not depend on which game are we
managing mods for, which is exactly what this change addresses.