Check connection to Thunderstore CDNs on app startup
A list of predefined CDNs are tested, and the first one that is reached
will be stored internally by CdnProvider as the preferred one. Other
parts of the code that deal with CDNs can use the static helper methods
provided by the class.

If no CDN can be reached, an error modal is shown to the user about the
issue. Hopefully the provided suggestion will help some users.

Refs TS-2003