Make the data prefix optional in profile imports
Make it possible to import base64 encoded profiles which don't have the
data prefix as the data prefix does not provide much practical value but
it does make debugging profile issues more difficult as decoding them
can't be done as easily with 3rd party tooling.

This is a forward looking change and does not impact anything in the
immediate future. Once this change has been rolled out, it's possible to
remove the prefix from the profile generation logic.