Use caching to improve performance of large profiles
When the local mod list is browsed vue seems to render the whole list
again on some user actions. This causes three methods of the component
to get called repeatedly:

- getMissingDependencies, which uses modifiableModList property, which
  is based on the mod list stored in VueX store. modifiableModList is
  used by other parts of the component as well, so I thought better not
  touch it at this time since figuring out unintended side effects
  would be a lot of work
- getThunderstoreModFromMod, which used the Thunderstore mod list
  stored in VueX, but now uses the list stored in the
  ThunderstorePackages. As far as I can tell both are updated in the
  Splash view and by the scheduled background process in UtilityMixin.
  Ergo this shouldn't break things, but this is the most significant
  functional change in this commit, and therefore most likely culprit
  should problems arise
- isLatestVersion, which did and still does use ThunderstorePackages
  for its shenanigans

So while this commit doesn't reduce the incessant function calls, it
caches the results to a simple object to reduce required calculations.
Effects were tested with a profile containing a mod pack with 109 mods.
Completing the following tasks were timed (roughly and manually), with
the accompanying results (original vs. cached):

- Initial rendering of the local mod list when moving from profile
  selection view: 7.0s vs. 5.4s
- Opening modal to disable a mod with two dependants: 4.4s vs. 1,2s
- Closing the modal without disabling the mod: 4.4s vs. 1.2s
- Opening modal to uninstall the same mod: 4.5s vs. 1.0s
- Uninstalling the mod: 15.8s vs 6.0s

(There might be further changes for optimizing the uninstall process,
since it seems some stuff is done after each dependant is uninstalled,
while it MIGHT be enough to do it just once in the end.)

For a small profile with 3 mods there's no noticeable difference
between the performance of the old and new implementation.