Postpone updating local mod list to improve uninstall performance
There's three methods that control how mods are uninstalled:
uninstallModRequireConfirmation, which calls performUninstallMod
directly if the mod has no dependants, and uninstallMod if the mod
has dependencies. uninstallMod in turn calls performUninstallMod for
the mod and all the dependants separately. This commit does the
following changes:

- By default, performUninstallMod now calls updateModListAfterChange
  helper, which differs from the original implementation in that it
  also calls filterModList method
- Since uninstallModRequireConfirmation calls performUninstallMod, it
  no longer needs to call filterModList directly
- performUninstallMod also accepts a new boolean parameter which can be
  used to prevent it from calling updateModListAfterChange
- uninstallMod uses the said parameter to prevent the
  updateModListAfterChange getting called after each individual mod is
  uninstalled. It calls updateModListAfterChange itself directly in the
  end

As far as I can tell this has no ill effects to the end result, i.e.
the profile files that remain on the disk appear to be identical. The
only downside I'm aware of is that the modal that shows the list of
mods no longer gets updated during the process, so it might appear to
the user that nothing happens.

The benefit is that the process takes significantly less time when
uninstalling a mod with lots of dependants. Prior to changes
uninstalling BepInEx from a profile that contained 109 mods took 147
seconds, while after the changes it "only" takes 24 seconds.