Disable package list updates on game selection and splash screens
Previously the package list was loaded for the default game while the
user was in the game selection screen. Now that so many games are
supported, the odds are that downloading RoR2 package list is a waste
of bandwidth.

Since the splash screen actively downloads the package list for the
selected game, triggering the background download would only slow
down both.