Create dedicated components for mod modals
Create dedicated components for mod management modals (uninstall,
disable, view assocaited). This commit does not move the state
management anywhere yet, but ideally it too would be separated from the
LocalModList.vue component.

While this commit does increase the amount of boilerplate in the
interim, it makes it easier to understand the state flow of the modals
(especially inside the modal component itself) and leaves room for
future changes that don't directly impact all of the modals.