Flush local mod list before moving to profile
LocalModList is rendered before transition to the view actually takes
place. If user has previously viewed a large profile, it's still stored
in VueX and this unnecessary step can take several seconds. Resetting
the stored mod list prevents this unnecessary wait.

Ideally this would be done when the mod list component unmounts, but
that would mess with the NavigationBar which shows the number of
installed mods.