Select installer using string IDs instead of enum
Convert the mod loader installer selection code to be slightly more data
oriented by using string IDs for package installers instead of relying
on a switch statement.

The main point of this change is to take a small step towards a design
where each package gets to declare which installers they want to be
processed by.

This is still equally well validated as the prior implementation thanks
to TypeScript.