Move install rules logic to InstallRuleInstaller
Move the install rule system installation logic to the new
InstallruleInstaller module.

The GenericProfileInstaller class will now instantiate an
InstallRuleInstaller object with the context-aware install rules applied
to it on creation.

Each install rule should be further broken down to a parametric install
rule handler, but for now things are kept as close to the original as
possible to make diffing easier. The main change in this commit was
turning the functions to module level functions instead of class
functions. This is a change that helps with future refactoring.