Move where manager sets doorstop target
Doorstop target was initially set in the mod list getter so the value
would get set in a new profile when BepInEx was installed. While this
works, it slows down the getter, which is called whenever the mod list
is updated, which happens quite often as user interacts with the local
mod list.

Instead, set the doorstop target when the modal showing the target is
opened. As far as I can tell, the doorstop is used here purely for
display reasons. The methods called are getters without side-effects,
and the GameRunner will call one or another when the game is launched
(choosing one based on whether the game is ran as modded or as
vanilla).