Fix race condition when setting doorstop
Use GameInstruction to only set vanillaLaunchArgs, and to not touch the
doorstopTarget, which should be set by the GameRunningProvider. This
way the doorstopTarget value shown on modal will actually match the
values that are used when the game is launched either as vanilla or as
modded by the GameRunner instance.