Add DeferredInput component
This is a simple input that stores the value internally and emits
debounced events when the value changes. The idea is that the value of
the input updates instanstly so it feels responsive, but heavier update
operations done by the parent component can be triggered only once the
debounce delay has expired.