Focus on text inputs when profile management modals are opened
nextTick method is used to ensure that the modals which us v-if are
mounted before attempting to focus on them.

I would've liked to also support using enter key to trigger the
action, but due to spaghetty modal reusage, that would require more
refactoring effort than I'm willing to invest at this point.