Remove state tracking from shimloader
As the stat tracking isn't generally used for the core mod loader
installers, remove it for shimloader as well. This also eliminates the
need for a handful of the workarounds that were in place to exempt a
specific file from being treated as any other file normally would for
state tracking.

This does not yet solve the initial "config should be placed in the game
dir" problem; solution for that upcoming as a followup.