Store active game and settings in Vuex store
Set the game when user proceeds (or is automatically redirected) from
game selection screen. Update the settings whenever game is changed
properly via action. To ensure the settings are properly loaded, they
should be accessed via getter rather than directly from the state. This
is not ideal but we're constrained by the fact that actions are the
only part of Vuex that supports async operations and provides access to
both local and root state/getters/mutations/actions.

The change means that the GameManager is no longer the single source of
truth in the matter. This is a trade-off with eventually being able to
do more business logic in Vuex store instead of the UI components.