Remove code that's never executed
The if-branch checks a variable that has been set to false just two
lines ago. I'm not sure if this is a bug that should be handled some
other way, but at least this part of the code has remained the same for
four years, so probably it's not very critical bug at least.