Move resetting mod list back to Profiles.vue
Commit b17d878 moved this bit of code from Profiles.vue to Manager.vue
since it makes for the latter to handle the state manipulation that
affects its own contents.

However, Vue doesn't await beforeCreate(), which opened up a race
condition where it's technically possible for the updating of the mod
list with an empty array to complete last, causing the view to render
an empty state.

While placing this bit of code in Profiles.vue is not ideal, at least
it works since the method is awaited there.