Limit the size of the profiles exported as a code to 20MB
If the zipped payload of the profile exceeds the limit, the process is
cancelled before the B64 encoding and upload steps, and an error modal
is shown to the user.

The size is limited firstly because there shouldn't be need for larger
profiles, and exceeding the limit is a symptom of something being
wrong, e.g. unintended contents in the BepInEx config folder.

Second reason is that too large profiles will crash on the B64 encoding
step on TSMM, which uses different file operation implementation than
r2mm.