Refactor how online mod list is updated in background
Originally the code first updated the list of available mods from
Thunderstore API and then loaded the local mod list from disk. In
commit dee959f yours truly (probably by mistake) changed this so that
both operations are executed independently on 5 minute intervals. this
probably leads to situations where the local mod list is loaded before
the online mod list is updated, which makes no sense, so revert to do
these operations in sequence.

(IDK why we load the mod list from the disk here in the first place but
that's outside the scope of this commit.)

Since refreshThunderstoreModList now also loads the mod list from disk,
this doesn't need to be separately handled in Manager/Settings view.

Loading the profile's mod list from disk was moved to Vuex action.
refreshThunderstoreModList could also be refactored into Vuex but since
it's used elsewhere too, it's not done within this commit.