Add PackageDexieStore
This IndexedDB wrapper is used for storing the responses received when
querying Thunderstore API for package listings for a specific game.

The refactoring shouldn't be visible to users (except for a minor delay
when storing the response) for now. In short term this allows us to
drop the duplicate in-memory list of ThunderstoreMod objects. In medium
term this aims to make it easier to do further refactorings to deal
with the poor performance of handling the very large mod lists in
communities like Lethal Company.