Fix deprecation displays
Use precalculated deprecation map that takes into account dependency
deprecations on LocalModCard.

Use the same map on OnlineModView to filter out deprecated mods unless
user uses filters to specifically ask for them.

And finally use the same map, via Vuex getter, to show the number of
mods available online on the navigation sidebar. This works a bit oddly
since the filter for deprecated mods is taken into account, but other
filters (e.g. NSFW content) are not. However, this is how it worked
before the changes too, and further changes are considered outside the
current scope.