Store timestamp of when the package list was last updated from API
Storing the value on IndexedDB is done in the same transaction as
storing the package list to ensure the values in DB stay in sync.

Reading the value into Vuex is done in the same action as reading the
package list in the hopes that they too stay in sync. There's nothing
to stop someone unknowingly updating the values separately through the
mutators though.