Show IndexedDB-related operations on Splash screen
First half of the progress is for storing the response to IndexedDB (if
a response was received) and the second half for reading the values
from database and casting to ThunderstoreMod objects to be stored in
Vuex.

Add a helper for resetting progress of all subitems on the splash
screen.