Prewarm mod cache when profile is selected
Changes that moved the mod cache from ModBridge class to Vuex caused
the initial rendering of the local mod list take roughly twice as long
as before (although subsequent renderings were a bit faster). Since
effort has been placed recently to improve the performance, this isn't
ideal.

One solution to improve the performance is to start prewarming the
cache as soon as a profile is selected. This obviously causes a slight
performance hit when a larger profile is selected, but overall this is
likely less noticeable to the user.

Once the cache is warmed, there's no point in resetting the mod list
when moving to Manager screen and reloading it there. These changes
will make moving to the Manager screen actually faster than it was
before (after the cache warming is done), and as a bonus the waiting
happens on profile screen, meaning the "no mods here" disclaimer does
not flash on screen while the mod list is loaded.