Prevent multiple ".old" suffixes when toggling mod disabled state
When mod disabled status was toggled multiple times while the first
change was processed, the files ended up with multiple suffixes. When
the mod was then toggled to enabled the files still had the suffix and
the mod would now be loaded.

The caveat is that any files that actually using .old as file extension
will get it ripped off if the mod is disabled and enabled again. While
not ideal, I see this as the lesser evil.