Unfix getDeprecatedPackageMap
It was deemed that simply fixing the method would be too drastic
change. We want the UI to be able to reflect the deprecation status
without fully hiding mods based on deprecated dependencies, and add
support for showing various other states like "verified" while we're at
it.

Meanwhile, intentionally break the new implementation in a manner that
is easily reverted once the UI changes are ready.