Improve mod lookup performance when importing profiles
Fetch only the mods the profile contains from IndexedDB, instead of all
the mods for the community. This is significantly faster for larger
communities like Lethal Company (roughly 80ms vs 2s). For very small
communities it might be a tiny bit slower, but the difference is
neglible.

It's worth noting that the IndexedDB lookup only checks the packages
from active game's community, which is in line with the old
implementation. I wasn't sure if it would make more sense to check if
the requested packages are available any of the communities. IndexedDB
contains only the mods from the previous queries though, so actually
querying all communities (or unlisted packages) would require a new
API queries. Furthermore, querying the IndexedDB by package name sans
community would require a new index to be created.