Prevent running multiple mod list background updates simultaneously
Online mod list is currently updated on the background every 5 minutes.
Ideally downloading and processing the mod list wouldn't take that long
that the interval would fire before an earlier round has finished, but
the most popular games have so many mods that on people with slower
connections this seems to happen. Multiple connections then compete for
the same download bandwidth, making the process even slower.