PackageDexieStore: write package list into IndexedDB in chunks
This is a speculative fix for an issue where at least one user receives
"Maximum IPC message size exceeded" error in this phase of the process
in SplashMixin.

The performance hit from this seem to mimimal, although the variance
seems high when it comes to timing IndexedDB operations. Something
similar might be done soon anyway when the package list fetching is
split into chunks too.