Add setting for toggling preferred Thunderstore CDN
For a long period of time, small percentage of users have been
encountering issues when downloading mods from Thunderstore CDN. First
attemt to remedy this was to create a checkup that runs when the mod
manager starts, and switches to using a CDN mirror.

However, sometimes users can ping the CDN health check endpoint on
startup and download most files, yet encounter issues with specific
files. Currently only solution has been to bypass the issue by using
another internet connection or a VPN, which are not available to all
users and are frankly pretty far from an ideal solution.

The new solution is to allow users to manually toggle the CDN
temporarily. The value is stored in memory and is lost when the app
closes. While this might seems like a shortcoming, it actually has
multiple benefits

- Thunderstore prefers the users to use the main CDN, so we don't
  want the users to turn on the mirror and then keep using it forever
  even when it's not needed
- This keeps things simple since we don't need to interact with the
  CDN health checkup when the app is started
- We also don't need to deal with Dexie storage schema migrations

The setting may become obsolete if the CDN ever starts working properly
but this is a small code change and easy to get rid off if need be.