Fix populating RenameProfileModal input with initial value
The previous implementation didn't take into account that the local
value is empty when user enters the Profiles screen, but an "old"
profile with a different name might be already selected if user did so
in another game's profile selection screen.

I think the proper solution would be to ensure that the active profile
is reset to null when user returns to game selection screen, but that's
a bigger change with a risk of breaking other stuff, so it's not
implemented now.