Change ReturnOfModding plugin uninstallation
Uninstalling a mod using ReturnofModding should remove the plugins_data
subdir. However, if it contains a cache subdir, the contents of that
subdir are kept, while rest of the content is removed.

This is done as a way to persist some of the data created by the plugin
when the mod is updated, since the update process uninstalls the old
version before the new one is installed. The behaviour is documented at
https://github.com/xiaoxiao921/ReturnOfModdingBase/tree/master?tab=readme-ov-file#plugins_data