Calculate content hashes when downloading blob files
The hashes will be used to identify previously seen package list
indices. The Thunderstore API uses the same hash to identify blobs.
The hash is recalculated on the mod manager because:

- The API returns a redirect response, and we can't access the headers
  of the original response, only the result of redirect from the CDN
- The redirect can't be prevented and handled manually in the browser
  environment
- The response from the CDN does contain the hash as its used in the
  file name, but the only place Axios stores it in the response is a
  field that looks like it could easily change in the future versions

Since hashing the amount of data we're handling should be in a scale of
single digit milliseconds per chunk, this shouldn't be a problem.