Skip unnecessary steps when downloading an unchanged mod list
In SplashMixin, skip downloading the chunks and updating them into the
IndexedDB if the index file indicates the packages haven't changed. If
downloading the index file itself fails, try to load earlier mod list
from the IndexedDB like it did earlier.

In UtilityMixin, skip the whole process if the index file indicates the
packages haven't changed, as at this point we already have the mod list
loaded in to Vuex. If loading the index file itself fails, the error is
caught by the caller, which handles retries as it did earlier.