Change where and how the "mods last updated" timestamp is updated
Old approach used to read the timestamp from the package list table
since it wasn't available elsewhere. This approach no longer works
when we don't update the package list into database if it didn't
change. Instead, store the information in the same table as the
community's last seen index chunk hash. This timestamp is updated
whenever the same index hash is seen again, or when the hash changes
AND the new package list is successfully written to DB.

Old approach loaded the timestamp into Vuex when the package list was
read to Vuex state. With the new approach we need to take into account
that UtilityMixin skips updating the package list in memory if the list
didn't change. Instead, we need to update the timestamp manually.