WIP Store target version of the mod loader of the active game in Vuex
Some of the communities share a common BepInEx package. If a newer
version is needed for one community, it's risky to just publish the
new version, since it would trigger the auto update feature for all
these communities, possibly breaking something. To prevent this, the
Thunderstore ecosystem schema will eventually include an optional
field for defining which version a community should use.

WIP: don't know where to read the value from yet.