Filter the mod list if a community targets a specific version of mod loader
Same BepInEx package can be listed in multiple communities. If a new
version is published for the package, we might want to target an older
version for some communities to avoid the chance of breaking things.

Thunderstore API returns all versions of the package. Filtering is done
on the list that is kept in memory by Vuex. In theory this approach
allows us to update the targeted version separately from reading the
whole mod list from the API and writing it to IndexedDB. It's not
necessarily what we end up doing, though.