Add ManagerUtils.filterModVersionsByDate
This helper can be used to filter mod versions by start and/or end
date. If all versions of a mod are filtered out, the whole mod is also
removed.

The idea is to create a "Time Bubble" feature. User could choose per
profile the dates between which the online mod list is filtered.
Intended use case is game updates breaking mods: if user updates the
game, they can filter out older versions to see only the mods that have
been updated after the breaking change. If user sticks to the old
version of the game, they can filter out newer mods that don't work on
their version, and avoid recommendations to update mods that would just
break things.