profileCreatedCallback: reduce nesting by removing setTimeout
I don't see any reason for to wrap the code in the timeout. This might
have been relevant when the import step was started right away after a
new profile was created, by there now exists a preview window between
these steps so the folder should always exists. If it doesn't, the
100ms wait won't change things.