Add new ImmutableProfile class that doesn't track the active profile
This is part of a larger refactoring effort aiming to remove the
business login from components so it can eventually be exported to be
a responsibility of a CLI tool.

Instead of inheritance, both classes only implement an interface
defining the common methods. This was done so that when a function
expects one of the classes, it won't accept the other. Being strict
about this makes it easier to see which part of the code base depend on
the active profile, and might reduce cases where passing wrong type of
object results in unexpected behaviour. The code duplication caused by
this within the classes seems like a small price to pay.

The static method for accessing the root dir of all profiles was
renamed to getRootDir() to make it clearer what directory it actually
returns. The method is not available on ImmutableProfile, since it
depends on the state of the app (active game) via PathResolver. We'll
likely add a non-static method to ImmutableProfile in a later commit
to reduce the dependency to app state where it's not required.